add history search

+++ src/main/java/com/warehouse/repository/InfoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InfoRepository extends JpaRepository<InfoEntity, Long> {
+public interface InfoRepository extends JpaRepository<InfoEntity, Integer> {
 
 }
