+++ src/main/java/com/warehouse/repository/QuantityOfProductRepository.java
@@ -4,6 +4,6 @@ import com.warehouse.domain.entity.QuantityOfProductEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @org.springframework.stereotype.Repository
-public interface QuantityOfProductRepository extends JpaRepository<QuantityOfProductEntity, Long> {
+public interface QuantityOfProductRepository extends JpaRepository<QuantityOfProductEntity, Integer> {
 
 }
